casa/requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
Deprecated==1.2.13
Django==4.0.5
djangorestframework==3.13.1
kombu==5.2.4
mysqlclient==2.1.0
packaging==21.3
Pillow==9.2.0
prompt-toolkit==3.0.30
pyparsing==3.0.9
pytz==2022.1
redis==4.3.4
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
